requirement.txt
fastapi==0.115.0
pymongo==4.9.1
python-dotenv==1.0.1
uvicorn==0.30.6